Well...
I added some more options to the quickstart menu (composer, tune type).

Then I was just about to stop, when I noticed a black box that appeared
over the first bar when I was inputting notes! Strange...

Turns out there's a bug in the vdom diffing that has been there for
about 9 months, that means if something is 0 (such as, in this case,
opacity) it is not drawn correctly. Somehow this didn't break until
today.

I also fixed another irritatingly obvious bug where when you move the
mouse to the previous bar, it doesn't actually move the preview note
over.