Make textbox dragging better
Before, it would always drag around the centre of the text box. This
meant that if you didn't click in the exact centre, the text box would
jump when you started dragging.

Now we store the offset when they click, so it doesn't matter where on
the textbox they click.