Use Mithril instead of custom virtual DOM diff
Using a more widely used VDOM library will result
in fewer bugs. Performance-wise, Mithril appears to
be a bit faster when no changes are necessary, and a
bit slower when they are.

This hasn't been pushed to 'production' yet - I'll
definitely need to check over corner cases. Doc help
is currently disabled.

In future, we could use Mithril's built-in autoreload
perhaps... not sure how that would work with dispatch()