Add a notice while loading audio samples
This greatly improves the experience on Safari, since it's now clear
what is happening while nothing is played.

These changes both make it clearer what is happening, and work more
reliably than the previous code, so that playback is now much less
flaky, especially on MacOS (touchwood - still needs a bit of testing).