Fix bug where things were dragged with the mouse up
Before now, selections were created to have start dragging immediately,
and stop at the first MouseUp event. This works if the selection is
created by a MouseDown event, but if it is created by a keyboard shortcut
(e.g. moveLeft) then it doesn't, since the mouse isn't down, and no MouseUp
event occurs, but dragging still happens.

Now every time a selection is created, a boolean must be passed in stating
whether or not the selection was created by a MouseDown event.