PipeScore: playback multiple second timings in one part correctly
Before, there was a bug where upon reaching the end of
the 2nd timing, it would set 'repeating' to false, and then
subsequent second timings would play the wrong section.

Now, we store the second timing over the final repeat mark,
and only set 'repeating' to false when we reach the end of that timing.