PipeScore: split up implementation/interface
I was running into circular dependency issues when writing a test
that imported Bar. This could be fixed by reordering imports, but
this is very fragile so hints at bad code :) this commit changes the
directory structure from:
Component/
	index.ts
	view.ts?

to:
Component/
	index.ts
	impl.ts
	view.ts

where index.ts contains an abstract class IComponent that can be imported
pretty much everywhere, impl contains the concrete class(es) and view
contains the view code for the abstract class.

Generally we want to keep it so that impl and view may both import
index but not the reverse. This keeps the interface code importable
from everywhere without circular dependency issues.

This fixes the circular dependency issue. One could argue it makes
the code cleaner (in that the interfaces are all well-defined).

There is still circular dependency in the code, but it's between
types which is OK.