Playback: clean up Sample code
Sample has become AudioResource, since that's really what it is.

Player has become Sample, likewise.

Drone and SoundedPitch now both use a Sample internally, so
they are on the same level of abstraction.

Sample and AudioResource are in their own file - these are the classes
wrapping the browser API.

Drone and SoundedPitch are in their own file - these are the classes
using Sample/AudioResource to play drones and pitches.

Documentation has also been added for these classes.