Playback: drastically improve harmony playback
Three major improvements!
- Fix 'drift' where harmonies gradually get more out of time
- Allow playback from/looping selection, even when there's harmonies
- Playback correctly when there are harmony staves missing

These were all accomplished by changing playback from working on
a simple list of notes, to working on a list of 'PlaybackMeasures',
where a PlaybackMeasure is just the notes from a measure.

This way:
- Indices (for start/end) no longer index into notes list, but
  into bars
- Indices also include a 'timeOffset' allowing starting/ending
  playing midway through a bar

Each measure is played before the next, synchronising the
different harmony parts at the barlines.

It also makes the code even more hard to understand - this has
been a problem in the playback code and it's not improving...
it's also not massively efficient, but this doesn't matter too
much since it's only run once at the start of playback, and
the drone is fading in then anyway so it's not noticeable -
I will test on a slower computer though before I deploy.

Incidental fixes:
- As a side-effect of a solution to ties over barlines, the cursor
  now updates for different notes in a tie
- Playback from/looping selection was only passing the first bar for
  some reason
- Bug fix in isRoughlyZero() :)